EduTech_Full/naplan-backend/requirements.txt
pymongo==4.6.3
python-dotenv==1.0.1
google-generativeai==0.7.2
langsmith==0.6.6
langchain-core
langchain-google-genai
langsmith
requests
Pillow
